Fix id type in repos

+++ src/main/java/ru/deloom/jpaservice/repositories/PagesRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import ru.deloom.jpaservice.entities.Pages;
 
-public interface PagesRepository extends JpaRepository<Pages, Integer>{
+public interface PagesRepository extends JpaRepository<Pages, Long>{
 	List<Pages> findByUrl(String url);
 }
